Adjusted changes to new document schema and fixed S3Key logic
